misc(form-file): update to BS 4.0.0
