+++ src/components/form-file/form-file.css
@@ -1,4 +1,3 @@
-<style>
     /* Custom-file focus styling: These can possibly be removed once BSV4.beta.3 is released */
     /* Although the focus handling/style will still be needed for Firefox keyboard-only users */
     /* Which will be addressed if PR https://github.com/twbs/bootstrap/pull/24138 is merged */
@@ -28,7 +27,7 @@
         box-shadow: 0 0 0 0.2rem rgba(40, 167, 69, 0.25);
         border-color: #28a745;
     }
-    
+
     /* Interim fix (until BS V4.beta.3) for is-{state} feedback for plain form-file input */
     /* See issue https://github.com/twbs/bootstrap/issues/24831 */
 
@@ -74,9 +73,3 @@
         color: white;
         content: attr(data-drop);
     }
-</style>
-
-<script>
-    import bFormFile from './_form-file'
-    export default bFormFile
-</script>
